requirements.txt
django==4.1.2
djangorestframework==3.14
openai==0.27.0